introduce leader_svc mod
